+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # WSU Accessibility Dashboard Changelog
 
+## 0.2.0 (August 29, 2017)
+
+* Display all domains in a sortable table on home page.
+* Calculate the errors per scanned URL on each domain.
+* Generate domain data with a node script.
+
 ## 0.1.4 (May 31, 2017)
 
 * Change deploy-to folder and bucket in S3.